module inception: move system/mod and system/system into system
